src/gausskernel/dbmind/tools/anomaly_detection/requirements.txt
python-dateutil
pandas
flask
flask_sqlalchemy
flask_restful
configparse
prettytable
pmdarima
